refactor: Extract AttendanceMemberInfo record from AttendanceRecordResponse
