+++ src/main/java/org/cotato/csquiz/api/attendance/dto/AttendanceRecordResponse.java
@@ -1,7 +1,6 @@
 package org.cotato.csquiz.api.attendance.dto;
 
 import org.cotato.csquiz.domain.auth.entity.Member;
-import org.cotato.csquiz.domain.auth.enums.MemberPosition;
 
 
 public record AttendanceRecordResponse(
@@ -13,19 +12,5 @@ public record AttendanceRecordResponse(
                 AttendanceMemberInfo.from(member),
                 attendanceStatistic
         );
-    }
-
-    public record AttendanceMemberInfo(
-            Long memberId,
-            String name,
-            MemberPosition position
-    ){
-        static AttendanceMemberInfo from(Member member) {
-            return new AttendanceMemberInfo(
-                    member.getId(),
-                    member.getName(),
-                    member.getPosition()
-            );
-        }
     }
 }
